+++ terminice/lib/src/selectors/choice_selector.dart
@@ -3,7 +3,6 @@ import 'dart:math';
 import 'package:terminice/terminice.dart';
 import 'package:terminice_core/terminice_core.dart';
 
-import '../core/component_runner.dart';
 import '../core/fallback_selection.dart';
 
 /// Extension providing the [choiceSelector] prompt for `Terminice`.